requirements-google-docs.txt
# Requirements for Google Docs API converter
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# Base requirements (from main requirements.txt)
lxml>=4.9.0
python-docx>=0.8.11
requests>=2.28.0 